Fix compatibility issues
