+++ src/components/index.ts
@@ -33,7 +33,6 @@ export * from './FileSelect';
 export * from './FileUploader';
 export * from './FileUploaderDetails';
 export * from './Flyout';
-export * from './Flyout2';
 export * from './Form';
 export * from './GradientContainer';
 export * from './Grid';
